Cleanup goalc tests, fix jak2 kernel bugs
