+++ test/goalc/test_control_statements.cpp
@@ -1,20 +1,10 @@
-#include <chrono>
-#include <cstdio>
-#include <iostream>
-#include <random>
-#include <sstream>
 #include <string>
 #include <thread>
 
-#include "inja.hpp"
-
 #include "game/runtime.h"
 #include "goalc/compiler/Compiler.h"
-#include "goalc/listener/Listener.h"
 #include "gtest/gtest.h"
 #include "test/goalc/framework/test_runner.h"
-
-#include "third-party/json.hpp"
 
 struct ControlStatementParam {
   // TODO - Not Needed Yet
@@ -23,7 +13,7 @@ struct ControlStatementParam {
 class ControlStatementTests : public testing::TestWithParam<ControlStatementParam> {
  public:
   static void SetUpTestSuite() {
-    runtime_thread = std::make_unique<std::thread>(std::thread((GoalTest::runtime_no_kernel)));
+    runtime_thread = std::make_unique<std::thread>(std::thread(GoalTest::runtime_no_kernel_jak1));
     compiler = std::make_unique<Compiler>(GameVersion::Jak1);
     runner = std::make_unique<GoalTest::CompilerTestRunner>();
     runner->c = compiler.get();
@@ -50,8 +40,6 @@ class ControlStatementTests : public testing::TestWithParam<ControlStatementPara
   static std::unique_ptr<GoalTest::CompilerTestRunner> runner;
 
   std::string testCategory = "control-statements";
-  inja::Environment env{GoalTest::getTemplateDir(testCategory),
-                        GoalTest::getGeneratedDir(testCategory)};
 };
 
 std::unique_ptr<std::thread> ControlStatementTests::runtime_thread;
@@ -59,50 +47,50 @@ std::unique_ptr<Compiler> ControlStatementTests::compiler;
 std::unique_ptr<GoalTest::CompilerTestRunner> ControlStatementTests::runner;
 
 TEST_F(ControlStatementTests, ConditionalCompilation) {
-  runner->run_static_test(env, testCategory, "conditional-compilation.static.gc", {"3\n"});
+  runner->run_static_test(testCategory, "conditional-compilation.static.gc", {"3\n"});
 }
 
 TEST_F(ControlStatementTests, Blocks) {
-  runner->run_static_test(env, testCategory, "nested-blocks-1.static.gc", {"7\n"});
-  runner->run_static_test(env, testCategory, "nested-blocks-2.static.gc", {"8\n"});
-  runner->run_static_test(env, testCategory, "nested-blocks-3.static.gc", {"7\n"});
+  runner->run_static_test(testCategory, "nested-blocks-1.static.gc", {"7\n"});
+  runner->run_static_test(testCategory, "nested-blocks-2.static.gc", {"8\n"});
+  runner->run_static_test(testCategory, "nested-blocks-3.static.gc", {"7\n"});
 }
 
 TEST_F(ControlStatementTests, GoTo) {
-  runner->run_static_test(env, testCategory, "goto.static.gc", {"3\n"});
+  runner->run_static_test(testCategory, "goto.static.gc", {"3\n"});
 }
 
 TEST_F(ControlStatementTests, Branch) {
-  runner->run_static_test(env, testCategory, "return-value-of-if.static.gc", {"123\n"});
+  runner->run_static_test(testCategory, "return-value-of-if.static.gc", {"123\n"});
 }
 
 TEST_F(ControlStatementTests, DoTimes) {
-  runner->run_static_test(env, testCategory, "dotimes.static.gc", {"4950\n"});
+  runner->run_static_test(testCategory, "dotimes.static.gc", {"4950\n"});
 }
 
 TEST_F(ControlStatementTests, Factorial) {
-  runner->run_static_test(env, testCategory, "factorial-recursive.static.gc", {"3628800\n"});
-  runner->run_static_test(env, testCategory, "factorial-iterative.static.gc", {"3628800\n"});
+  runner->run_static_test(testCategory, "factorial-recursive.static.gc", {"3628800\n"});
+  runner->run_static_test(testCategory, "factorial-iterative.static.gc", {"3628800\n"});
 }
 
 TEST_F(ControlStatementTests, Definitions) {
-  runner->run_static_test(env, testCategory, "defun-return-constant.static.gc", {"12\n"});
-  runner->run_static_test(env, testCategory, "defun-return-symbol.static.gc", {"42\n"});
+  runner->run_static_test(testCategory, "defun-return-constant.static.gc", {"12\n"});
+  runner->run_static_test(testCategory, "defun-return-symbol.static.gc", {"42\n"});
 }
 
 TEST_F(ControlStatementTests, ReturnValue) {
-  runner->run_static_test(env, testCategory, "return.static.gc", {"77\n"});
-  runner->run_static_test(env, testCategory, "return-arg.static.gc", {"23\n"});
-  runner->run_static_test(env, testCategory, "return-colors.static.gc", {"77\n"});
+  runner->run_static_test(testCategory, "return.static.gc", {"77\n"});
+  runner->run_static_test(testCategory, "return-arg.static.gc", {"23\n"});
+  runner->run_static_test(testCategory, "return-colors.static.gc", {"77\n"});
 }
 
 TEST_F(ControlStatementTests, Calling) {
-  runner->run_static_test(env, testCategory, "nested-call.static.gc", {"2\n"});
-  runner->run_static_test(env, testCategory, "simple-call.static.gc", {"30\n"});
+  runner->run_static_test(testCategory, "nested-call.static.gc", {"2\n"});
+  runner->run_static_test(testCategory, "simple-call.static.gc", {"30\n"});
 }
 
 TEST_F(ControlStatementTests, Anonymous) {
-  runner->run_static_test(env, testCategory, "lambda-1.static.gc", {"2\n"});
+  runner->run_static_test(testCategory, "lambda-1.static.gc", {"2\n"});
 }
 
 TEST_F(ControlStatementTests, InlineIsInline) {
@@ -121,7 +109,7 @@ TEST_F(ControlStatementTests, InlineIsInline) {
     }
   }
   EXPECT_TRUE(got_mult);
-  runner->run_static_test(env, testCategory, "declare-inline.static.gc", {"32\n"});
+  runner->run_static_test(testCategory, "declare-inline.static.gc", {"32\n"});
 }
 
 TEST_F(ControlStatementTests, AllowInline) {
@@ -143,54 +131,54 @@ TEST_F(ControlStatementTests, AllowInline) {
   }
   EXPECT_EQ(got_mult, 1);
   EXPECT_EQ(got_call, 1);
-  runner->run_static_test(env, testCategory, "inline-call.static.gc", {"44\n"});
+  runner->run_static_test(testCategory, "inline-call.static.gc", {"44\n"});
 }
 
 TEST_F(ControlStatementTests, ReturnNone) {
-  runner->run_static_test(env, testCategory, "function-returning-none.static.gc", {"1\n"});
+  runner->run_static_test(testCategory, "function-returning-none.static.gc", {"1\n"});
 }
 
 TEST_F(ControlStatementTests, InlineBlock1) {
-  runner->run_static_test(env, testCategory, "inline-with-block-1.static.gc", {"1\n"});
+  runner->run_static_test(testCategory, "inline-with-block-1.static.gc", {"1\n"});
 }
 
 TEST_F(ControlStatementTests, InlineBlock2) {
-  runner->run_static_test(env, testCategory, "inline-with-block-2.static.gc", {"3\n"});
+  runner->run_static_test(testCategory, "inline-with-block-2.static.gc", {"3\n"});
 }
 
 TEST_F(ControlStatementTests, InlineBlock3) {
-  runner->run_static_test(env, testCategory, "inline-with-block-3.static.gc", {"4\n"});
+  runner->run_static_test(testCategory, "inline-with-block-3.static.gc", {"4\n"});
 }
 
 TEST_F(ControlStatementTests, InlineBlock4) {
-  runner->run_static_test(env, testCategory, "inline-with-block-4.static.gc", {"3.0000\n0\n"});
+  runner->run_static_test(testCategory, "inline-with-block-4.static.gc", {"3.0000\n0\n"});
 }
 
 TEST_F(ControlStatementTests, ReturnFromTrick) {
-  runner->run_static_test(env, testCategory, "return-from-trick.static.gc", {"1\n"});
+  runner->run_static_test(testCategory, "return-from-trick.static.gc", {"1\n"});
 }
 
 TEST_F(ControlStatementTests, Set) {
-  runner->run_static_test(env, testCategory, "set-symbol.static.gc", {"22\n"});
+  runner->run_static_test(testCategory, "set-symbol.static.gc", {"22\n"});
 }
 
 TEST_F(ControlStatementTests, Protect) {
-  runner->run_static_test(env, testCategory, "protect.static.gc", {"33\n"});
+  runner->run_static_test(testCategory, "protect.static.gc", {"33\n"});
 }
 
 TEST_F(ControlStatementTests, Align) {
-  runner->run_static_test(env, testCategory, "align16-1.static.gc", {"80\n"});
-  runner->run_static_test(env, testCategory, "align16-2.static.gc", {"64\n"});
+  runner->run_static_test(testCategory, "align16-1.static.gc", {"80\n"});
+  runner->run_static_test(testCategory, "align16-2.static.gc", {"64\n"});
 }
 
 TEST_F(ControlStatementTests, Defsmacro) {
-  runner->run_static_test(env, testCategory, "defsmacro-defgmacro.static.gc", {"20\n"});
+  runner->run_static_test(testCategory, "defsmacro-defgmacro.static.gc", {"20\n"});
 }
 
 TEST_F(ControlStatementTests, Desfun) {
-  runner->run_static_test(env, testCategory, "desfun.static.gc", {"4\n"});
+  runner->run_static_test(testCategory, "desfun.static.gc", {"4\n"});
 }
 
 TEST_F(ControlStatementTests, DeReference) {
-  runner->run_static_test(env, testCategory, "methods.static.gc", {"#t#t\n0\n"});
+  runner->run_static_test(testCategory, "methods.static.gc", {"#t#t\n0\n"});
 }
